requirements.txt
arrow==0.15.2
asgiref==3.2.3
autopep8==1.4.4
beautifulsoup4==4.8.0
binaryornot==0.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
DateTime==4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-admin-sortable2==0.7.3
django-allauth==0.41.0
django-angular==2.2.1
django-appconf==1.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==1.0.1
django-classy-tags==0.9.0
django-cms==3.5.3
django-cors-headers==3.2.1
django-filer==1.5.0
django-filter==2.2.0
django-formtools==2.1
django-fsm==2.6.1
django-fsm-admin==1.2.4
django-ipware==2.1.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-post-office==3.2.1
django-rest-auth==0.9.5
django-sekizai==1.0.0
django-select2==6.3.1
django-shop==1.0.2
django-tables2==2.1.1
django-treebeard==4.3
django-widget-tweaks==1.4.5
djangocms-admin-style==1.4.0
djangocms-cascade==0.19
djangocms-text-ckeditor==3.8.0
djangorestframework==3.10.3
easy-thumbnails==2.6
feedparser==5.2.1
future==0.17.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
Jinja2==2.10.1
jinja2-time==0.2.0
jsonfield==2.0.2
lxml==4.3.4
MarkupSafe==1.1.1
oauthlib==3.1.0
pathlib==1.0.1
Pillow==6.1.0
pipenv==2018.11.26
poyo==0.5.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
Unidecode==1.0.23
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
webencodings==0.5.1
whichcraft==0.6.1
xmltodict==0.12.0
zope.interface==4.6.0
